+++ tests/Helpers/AddonFiles.lua
@@ -29,6 +29,7 @@ M.migrator = {
 	"src/Config/Migrations/V74.lua",
 	"src/Config/Migrations/V79.lua",
 	"src/Config/Migrations/V80.lua",
+	"src/Config/Migrations/V81.lua",
 }
 
 function M.load(files, addon)